New files uploaded
Two different fucntions that are still under improvement, Fit channels gives back the alpha an beta from an array with current values from single channel patch clamp. The IVplotConductance gives the conductance plot and makes linear regression to the IV curve from 0 to 40 voltage. Then calculates reversal potential and conductance, and plots it. The input is a csv file with voltage in the forst column and current in the second column.